# Improve khmer segmenter performance by using fst segmenter