… charabia/dictionaries/fst/khmer/words.fst