# Pull Request

## Related issue
Fixes [#250](https://github.com/meilisearch/charabia/issues/250)

## What does this PR do?
- [x] add khmer words fst converted from [khmerdict.txt](https://github.com/unicode-org/icu/blob/main/icu4c/source/data/brkitr/dictionaries/khmerdict.txt) using `fst` crate
- [x] change to use fst segmenter instead of `icu_segmenter`

## PR checklist
Please check if your PR fulfills the following requirements:
- [x] Does this PR fix an existing issue, or have you listed the changes applied in the PR description (and why they are needed)?
- [x] Have you read the contributing guidelines?
- [x] Have you made sure that the title is accurate and descriptive of the changes?

Thank you so much for contributing to Meilisearch!