@ManyTheFish 2023-11-23
Hello @xshadowlegendx,
thank you for your PR,
did you run the benchmarks? And did you see any improvements? 😄